PR-SUG: extract inner-class in TestApiServer to configure the TLS client certificate
